+++ src/main/java/nextstep/subway/dto/StationRequest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.dto;
 
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.station.Station;
 
 public class StationRequest {
     private String name;
