+++ src/test/java/nextstep/subway/domain/StationTest.java
@@ -3,6 +3,8 @@ package nextstep.subway.domain;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import nextstep.subway.domain.station.Station;
+import nextstep.subway.domain.station.StationName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
